Read and write file text

+++ 0x15-file_io/0-read_textfile.c
@@ -10,7 +10,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	ssize_t file, txt;
+	int file, txt;
 	char *buffer;
 
 	if (filename == NULL)
